requirements.txt
playwright
tinydb
colorlog
python-dotenv
win10toast_click
pyperclip